Decisions Implemented
I added a game manager that controls lives, balls, and creating a new
paddle when you lose a ball. I also added a scene in between each level
that allows the user to choose another level, return to their home
planet, or the main menu. If they have completed a level, that button is
disabled.